requirements.txt
fastapi
uvicorn
faker
psycopg2-binary
python-dotenv
